refactor repository tests
